Reduce a dbus queries task count to '1'.
Currently, the DBus queries are processing at the '10' threads.
Since the current implementation of watchers not thread-safe, the
connection opened for each thread task.
The 'massif' analyzer shows that every one sdbusplus connection placed
a very cost of memory.

```
->51.34% 0x1329AF: createDbusConnection() (dbus_broker.cpp:310)
|->34.23% 0x14B114: doWatchDBusServiceOwners() (dbus_broker.cpp:200)
```

Hence, the reducing count of query tasks will reduce the count of active
sdbusplus connections.  Besides, the dbus-query is not related to the
http-request, and, in fact, the direct dbus query processing only one
time per 'yaweb' process-life.

End-user-impact: The total memory usage by the 'yaweb' process was
                 reduced by ~80%

Signed-off-by: Igor Kononenko <i.kononenko@yadro.com>