Query: Fix extract complex primitives.
Currently, the list of strings and doubles does extract as complex
dbus-instances and expands to the parent entity. That is incorrect and
following to impossible to retrieve the dbus array-like types (except
associations).
This patch brings support of retrieve array strings and doubles from
dbus-query and pass-through to router-handler through IEntity.

End-user-impact: None.

Signed-off-by: Igor Kononenko <i.kononenko@yadro.com>