IEntity: Add a Type of IEntity.
The current implementation has Entity abstraction that contains
instances of specified type. The Entity might hold several IInstance
objects and the EntityManager client desire to know how to interpret
that item set. This patch brings Type of IEntity:
 * Object: IEntity contains explicitly one IInstance.
 * Collection: IEntity might contains a number objects of IInstance.
 * Reference: IEntity contains a reference to anather IEntity object.

Based on this enhancement the GQL can explicitly determine the type of
requested node: whether a json-object or a json-array.

End-user-impact: None.

Signed-off-by: Igor Kononenko <i.kononenko@yadro.com>