redfish: Add a way to pretty print response for text/html
There is a way to determinate that request incoming from browser by
Accept header that indicates the text/html content-type is prefered.

The such request would be better to answer pretty printed redfish
payload. The current patch brings that way.

End-user-impact: Now the browser requests to the redfish enpointes will
                 be pretty-formated to the html view.

Signed-off-by: Igor Kononenko <i.kononenko@yadro.com>