app: Provide webserver initialization progress.
The initialization of the yaweb depends on the full system state, and
after internal cache will be initialized a number of important
abstractions, e.g. Server, Chassis, etc. may be hasn't relevant objects,
because the  EntityManager initializion get some time.

This patch brings a way to provide yaweb status by create lock-file on
start and remove it on the initialization is completed.  This will
helpfull for the another processes, e.g. bmcweb to determinate that the
yaweb was full-initialized.

End-user-impact: None (internal system feature)

Signed-off-by: Igor Kononenko <i.kononenko@yadro.com>