dbus-query:cache:fix dbus-path depth
When handling the dbus-signals, the IInstance cache is updated,
and the DBusQuery endpoint criteria determine the purpose of such
dbus-signals. There is an error in the method of calculation of
the object path depth in relation to the dbus-namespace of the criteria.

This patch fixes that.

End-user-impact: Redfish: fixed bugs in object caching

Signed-off-by: Igor Kononenko <i.kononenko@yadro.com>
Signed-off-by: Alexander Filippov <a.filippov@yadro.com>