dbus-query:cache: Fix handling interface Added/Removed signals
There is an engine that observes dbus-signals InterfacesAdded and
InterfacesRemoved.  Currently, the engine registers a handler for each
DBusQuery. Each such a handler is registerd with the same set of
signal-catching rules. Matching of a signal to the corresponding
DBusInstance is performed in the callback-handler, and that works
without errors. Nonetheless, in the described architecture there are too
many identical signals that are registered in dbus-broker imposing
a serious overhead to DBus that significantly slows down signal
processing.

The patch aims to change the way of how the dbus-signal handlers are
registered. In dbus-broker there will be only one signal handler per
InterfacesAdded / InterfacesRemoved dbus-signal, plus some additional
internal logic in the handler-callback to match the signal to
a DbusQuery.

Ens-user-impact: WebUI/Redfish: Improved (shortened) response time, more
                 up-to-date information

Signed-off-by: Igor Kononenko <i.kononenko@yadro.com>