redfish: Fix format of anchor expanded objects
Redfish specification defines a way to identify objects that
are expanded internaly on a Node. DMTF requires that there should be
no slash in at the end of a URI before an anchor.

This patch fixes the issue according to that requirement.

End-user-impact: Now Redfish provides valid ODataID field for expanded
                 objects.

Signed-off-by: Igor Kononenko <i.kononenko@yadro.com>