core: fix: avoid excessive dbus requests
The DBusQuery::createInstance() sends requests to DBus to fill the
Entity fields with the specified properties that are provided by the
IQuery endpoint configuration. The set of interfaces that will be
introspected is asked by GetSubTree DBus method and contains
all interfaces of the target dbus object.

The patch modifies the logic and makes the code only do dbus calls
for the specified interfaces in the IQuery endpoint configuration.
That approach reduces time to create an entire IInstance.

End-user-impact: Now initializing DBus cache takes
                 a little bit less time.

Signed-off-by: Igor Kononenko <i.kononenko@yadro.com>