core: entity: add a way to invalidate cache
The dbus calls may fail and those cases aren't handled.
The patch brings an event mechanism that provides a way to invalidate
the IEntity cache when the corresponding IQuery fails.

If an IEntity cache is invalidated then another attempt to fill an
IEntity will be made on each access to it.

End-user-impact: Now GraphQL and Redfish provide consistent data only,
                 the irrelevant (dummy) nodes/objects/fields are no
                 longer reported.

Signed-off-by: Igor Kononenko <i.kononenko@yadro.com>