core: query: fix weak pointer reference
The RemovedInterfaces signal handler captures dbusQuery smart pointer,
that is a weak pointer and it does not hold the object throughout the
whole time of the signal processing. That may lead to segfault when
dbusQuery object is removed at the same time as the signal is handled.

End-user-impact: yaweb is more stable

Signed-off-by: Igor Kononenko <i.kononenko@yadro.com>