dbus:connect: increase timeout to watch dbus-signals
The dbus-watch thread has a very short interval between process_discard
calls - 1ms, that is lead to excessive overhead at the DBus at all.
At an other hand this is unnecessary, because the claimed accuracy is
not required.

The patch increases timeout between process_discard to the 20ms and
DBusWindowCall to the 500us (consequential change).

End-user-impact: Now dbus-requests that have direct calls to the dbus
                 handling a little bit quickly, the total cpu/memory
                 usage are decreased.

Signed-off-by: Igor Kononenko <i.kononenko@yadro.com>